requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
decorator==5.1.1
distlib==0.3.8
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docx==0.2.4
executing==2.0.1
Faker==25.2.0
filelock==3.14.0
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
ipython==8.13.0
jedi==0.19.1
lxml==5.2.1
matplotlib-inline==0.1.7
msgpack==1.0.8
packaging==24.0
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.1
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
redis==5.0.4
requests==2.31.0
rsa==4.9
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
tqdm==4.66.4
traitlets==5.14.3
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.26.1
wcwidth==0.2.13
whitenoise==6.6.0
wikipedia==1.4.0
Wikipedia-API==0.6.0
zope.interface==6.4
